requirements.txt
# ── Core ──────────────────────────────────────────────────────
flask==3.1.0
flask-cors==5.0.0
flask-jwt-extended==4.7.1
flask-limiter==3.9.3
python-dotenv==1.1.0

# ── Database ──────────────────────────────────────────────────
pymongo==4.11.3
dnspython==2.7.0

# ── AI / Gemini ───────────────────────────────────────────────
google-generativeai==0.8.4

# ── Production WSGI ───────────────────────────────────────────
gunicorn==23.0.0

# ── Security & Utils ──────────────────────────────────────────
cryptography==44.0.2
